requirements.txt
fastapi==0.125.0
uvicorn[standard]==0.38.0
requests==2.32.3
pydantic==2.12.3
python-multipart==0.0.20
python-dotenv==1.0.0

# Dependencias del Motor de Análisis
joblib==1.3.2
numpy
pandas
groq==0.13.0
sentence-transformers
scikit-learn
shap
